+++ src/subscription/TimeAware.sol
@@ -10,9 +10,8 @@ abstract contract TimeAware {
 
     /**
      * @notice the current time unit
-     * @dev the current time is represented as an i64. Higher values seem unreasonable
      * @return the current time unit
      */
-    function _now() internal view virtual returns (uint64);
+    function _now() internal view virtual returns (uint256);
 
 }
